Move file into assets folder
