+++ src/assets/js/script.js
